requirements.txt
# Canonical deps live in pyproject.toml [project.dependencies].
# Keep this file in sync for Docker layer caching.
python-telegram-bot[job-queue]==21.6
anthropic>=0.79.0
python-dotenv==1.0.1
requests>=2.31.0
ddgs>=7.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
aiohttp>=3.9.0
